fix api_tracker
Now it can hook apis from the library's plt entry. But not sure why the address not detected to be symbolic in libdummy test.

Probably need to reconsider the logic in analysis to make it step by step run and use the solver from the state. Or need to place the state memory to a completely random address to make it purely symbolic, no need to reference the concrete value.